+++ app/src/main/java/com/techcourse/service/UserService.java
@@ -1,43 +1,11 @@
 package com.techcourse.service;
 
-import com.techcourse.dao.UserDao;
-import com.techcourse.dao.UserHistoryDao;
 import com.techcourse.domain.User;
-import com.techcourse.domain.UserHistory;
-import org.springframework.jdbc.core.TransactionManager;
 
-public class UserService {
+public interface UserService {
 
-  private final UserDao userDao;
-  private final UserHistoryDao userHistoryDao;
+  User findById(final long id);
+  void insert(final User user);
+  void changePassword(final long id, final String newPassword, final String createBy);
 
-  public UserService(final UserDao userDao, final UserHistoryDao userHistoryDao) {
-    this.userDao = userDao;
-    this.userHistoryDao = userHistoryDao;
-  }
-
-  public User findById(final long id) {
-    return userDao.findById(id);
-  }
-
-  public void insert(final User user) {
-    userDao.insert(user);
-  }
-
-  public void changePassword(final long id, final String newPassword, final String createBy) {
-    final var user = findById(id);
-
-    TransactionManager.start();
-    try {
-      user.changePassword(newPassword);
-      userDao.update(user);
-      userHistoryDao.log(new UserHistory(user, createBy));
-      TransactionManager.commit();
-    } catch (Exception e) {
-      TransactionManager.rollback();
-      throw e;
-    } finally {
-      TransactionManager.release();
-    }
-  }
 }
